Add per-architecture transformer blocks with single-layer parity tests
Implement LlamaTransformerBlock (pre-norm, no QK-norm, SwiGLU),
Qwen3TransformerBlock (pre-norm, QK-norm, SwiGLU), and
Gemma3TransformerBlock (sandwich norm, QK-norm, GeGLU, (1+weight)
RMSNorm convention, query_pre_attn_scalar).

Fix RMSNorm casting order to match HF convention: cast normed back
to input dtype before weight multiply, achieving bit-identical parity
with HF reference models (zero error on Qwen3 and Gemma3).

+++ src/infer/models/common.py
@@ -32,7 +32,8 @@ class RMSNorm(nn.Module):
         input_dtype = x.dtype
         x = x.to(torch.float32)
         normed = x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + self.eps)
-        return (self.weight * normed).to(input_dtype)
+        # Cast normed back before weight multiply to match HF rounding convention.
+        return self.weight.to(input_dtype) * normed.to(input_dtype)
 
 
 # ---------------------------------------------------------------------------
